refactor: BoardService 패키지 변경
