+++ week04/kyulim/everytime/src/main/java/com/example/everytime/board/application/BoardService.java
@@ -1,4 +1,4 @@
-package com.example.everytime.board;
+package com.example.everytime.board.application;
 
 import com.example.everytime.board.api.dto.BoardListResDto;
 import com.example.everytime.board.api.dto.BoardResDto;
